Upgrade server: add async errorHandler, move routes to separate directory
