+++ frontend/src/config.js
@@ -9,7 +9,7 @@ const dev = process.env.NODE_ENV === `development`
 // TODO TEMP
 // const stargate = `https://lcd.nylira.net`
 // const stargate  = `https://us-central1-staking-explorer-c9991.cloudfunctions.net/mocks`
-const stargate = `http://localhost:8080/`
+const stargate = `http://localhost:8080`
 
 // const rpc = `https://rpc.nylira.net`
 const rpc = ``
